Updated .gitignore and modularized the code of trainer.py for easier modification in the future
