+++ src/data_loader.py
@@ -1,0 +1,33 @@
+import torch
+import torchvision.datasets as datasets
+import torchvision.transforms as transforms
+
+def get_data_loaders(batch_size, workers):
+
+    # Normalize the data
+    normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406],
+                                     std=[0.229, 0.224, 0.225])
+
+    # Return train_loader and val_loader
+    print("Setting up data loaders...")
+    train_loader = torch.utils.data.DataLoader(
+        datasets.CIFAR10(root='./data', train=True, transform=transforms.Compose([
+            transforms.RandomHorizontalFlip(),
+            transforms.RandomCrop(32, 4),
+            transforms.ToTensor(),
+            normalize,
+        ]), download=True),
+        batch_size=batch_size, shuffle=True,
+        num_workers=workers, pin_memory=True)
+
+    val_loader = torch.utils.data.DataLoader(
+        datasets.CIFAR10(root='./data', train=False, transform=transforms.Compose([
+            transforms.ToTensor(),
+            normalize,
+        ])),
+        batch_size=128, shuffle=False,
+        num_workers=workers, pin_memory=True)
+
+    print("Data loaders set up complete.")
+
+    return train_loader, val_loader
